erp2/requirements.txt
grpcio==1.56.2
grpcio-tools==1.56.2
pika==1.3.1
aio-pika==9.2.2
sqlalchemy==2.0.27
asyncpg==0.29.0
python-dotenv==1.0.0
asyncio==3.4.3